Memmory validation fix

+++ CodeInject/Form1.cs
@@ -14,16 +14,20 @@ namespace CodeInject
 {
     public partial class Form1 : Form
     {
-        string selectedMonster = "";
+        INPC selectedMonster ;
         PlayerCharacter player;
-        List<Monster> MonsterList = new List<Monster>();
+        List<INPC> MonsterList = new List<INPC>();
 
 
 
         public Form1()
         {
             InitializeComponent();
-            RefreshMonsters();
+            RefreshMonstersListBox();
+
+
+            RefreshMonstersList();
+
         }
 
 
@@ -40,14 +44,34 @@ namespace CodeInject
             player.Update();
         }
 
-        public void RefreshMonsters()
+        public void RefreshMonstersListBox()
         {
             listBox1.Items.Clear();
-            MonsterList.RemoveRange(0, MonsterList.Count);
+
+            ulong MonsterListAdr = GameMethods.GetInt64(GameMethods.GetBaseAdress() + 0x10C8410);
+            MonsterListAdr = GameMethods.GetInt64(MonsterListAdr + 0x22050);
+
+            MonsterListAdr += 0x4;
+            int monsterIndex = 0;
+
+            while ((monsterIndex = GameMethods.GetInt32(MonsterListAdr)) != 0)
+            {
+                listBox1.Items.Add(GameMethods.GetInt32(MonsterListAdr));
+                MonsterListAdr += 0x4;
+            }
+
+            label2.Text = $"Monster Count:{MonsterList.Count}";
+        }
+
+        public void RefreshMonstersList()
+        {
+            if (MonsterList.Count > 0)
+                MonsterList.RemoveRange(0, MonsterList.Count);
 
             ulong MonsterListAdr = GameMethods.GetInt64(GameMethods.GetBaseAdress() + 0x10C8410);
             MonsterListAdr = GameMethods.GetInt64(MonsterListAdr + 0x22050);
             LoadPlayerInformation(MonsterListAdr);
+
             MonsterListAdr += 0x4;
             int monsterIndex = 0;
 
@@ -55,56 +79,79 @@ namespace CodeInject
             {
                 Monster mob = new Monster(GameMethods.GetInt32(MonsterListAdr));
                 mob.Update();
-                listBox1.Items.Add(mob.ID.ToString());
+
                 MonsterList.Add(mob);
 
                 MonsterListAdr += 0x4;
             }
+
+            label2.Text = $"NPC Count:{MonsterList.Count}";
         }
 
 
         private void button2_Click(object sender, EventArgs e)
         {
             timer2.Enabled = !timer2.Enabled;
+            timer1.Enabled = !timer1.Enabled;
+
 
             button2.Text = timer2.Enabled == true ? "Stop" : "Start";
+
+            selectedMonster = null;
         }
 
 
         private void timer2_Tick(object sender, EventArgs e)
         {
-            RefreshMonsters();
+            if (selectedMonster != null)
+            {
+                Random rand = new Random();
+                GameMethods.AttackTarget((uint)rand.Next(0x210, 0x212), (uint)selectedMonster.ID);
+            }
+            // label1.Text = $"{selectedMonster} {isAlive} Distance:{Vector2.Distance(MonsterList.FirstOrDefault(x=>x.ID==int.Parse(selectedMonster)).Position, player.Position)}";
+        }
+
+        private void timer1_Tick(object sender, EventArgs e)
+        {
+            RefreshMonstersListBox();
+            RefreshMonstersList();
             if (listBox1.Items == null)
                 return;
 
-            bool isAlive = false; 
 
 
-            for(int i=0;i<listBox1.Items.Count;i++)
+
+            bool isAlive = false;
+
+            if (selectedMonster != null)
             {
-                if (listBox1.Items[i].ToString() == selectedMonster)
+                for (int i = 0; i < listBox1.Items.Count; i++)
                 {
-                    isAlive = true;
-                    break;
+                    if (listBox1.Items[i].ToString() == selectedMonster.ID.ToString())
+                    {
+                        isAlive = true;
+                        break;
+                    }
                 }
             }
 
-            if (isAlive)
-            {
-                Random rand = new Random();
-                GameMethods.AttackTarget((uint)rand.Next(0x210, 0x212), uint.Parse(selectedMonster));
-            }
-            else
+
+            if (!isAlive || selectedMonster == null)
             {
                 Random skillRandomizer = new Random();
-                Monster closestMonster = MonsterList.OrderBy(x => Vector2.Distance(x.Position, player.Position)).FirstOrDefault();
-                selectedMonster = closestMonster.ID.ToString();
-                GameMethods.AttackTarget((uint)skillRandomizer.Next(0x210, 0x212), uint.Parse(selectedMonster));
+                INPC closestMonster = MonsterList.OrderBy(x => Vector3.Distance(x.Position, player.Position)).FirstOrDefault();
+                selectedMonster = closestMonster;
+                Logger.Logger.Log($"New Attack Target: {selectedMonster}");
+                GameMethods.AttackTarget((uint)skillRandomizer.Next(0x210, 0x212), (uint)selectedMonster.ID);
             }
 
+     
 
-            label1.Text = $"{selectedMonster} {isAlive} Distance:{Vector2.Distance(MonsterList.FirstOrDefault(x=>x.ID==int.Parse(selectedMonster)).Position, player.Position)}";
+            lPlayerPosition.Text = $"Position: {player.Position.ToString()}";
+            lPlayerID.Text = $"Position: {player.ID}";
+
+            lNpcID.Text = $"ID: {selectedMonster.ID}";
+            lNpcPosition.Text = $"Position: {selectedMonster.Position}";
         }
-
     }
 }
